Update CMake configuration, fix data types in simulation, and enhance evaluation script
- Set TORCH_CUDA_ARCH_LIST to a fixed value in CMakeLists.txt.
- Change data type from uint8_t to uint16_t in simulate.cpp for better compatibility.
- Add a new evaluate_parallel.py script for batch evaluation with parallel Stockfish instances.
- Adjust replay buffer capacity in train_cpp.py and add a new bootstrap replay buffer capacity argument.
- Create a test script to compare tensor outputs between Python and C++ implementations.

+++ src/cpp/src/simulate.cpp
@@ -398,7 +398,7 @@ void simulate_games_buffered(
 
                     std::vector<float> flat;
                     flat.reserve(raw.size());
-                    for (uint8_t b : raw)
+                    for (uint16_t b : raw)
                         flat.push_back(static_cast<float>(b));      // 0.f / 1.f
 
                     auto policy = run_mcts(game, args);
